Refactor restart tests from mocha to tap
