+++ lib/master.js
@@ -6,7 +6,9 @@ var EventEmitter = require('events').EventEmitter;
 var _ = require('lodash');
 var assert = require('assert');
 var cluster = require('cluster');
+var clusterSize = require('./util').clusterSize;
 var debug = require('debug')('strong-cluster-control:master');
+var liveWorkerIds = require('./util').liveWorkerIds;
 var msg = require('./msg');
 var nextTick = setImmediate;
 var os = require('os');
@@ -14,22 +16,6 @@ var util = require('util');
 
 // Master is a singleton, as is the cluster master
 var master = new EventEmitter();
-
-// Workers remain in cluster.workers after they have have exited.
-function liveWorkerIds() {
-  var workerIds = Object.keys(cluster.workers).filter(function(id) {
-    var worker = cluster.workers[id];
-    var isAlive = worker.process.signalCode == null &&
-      worker.process.exitCode == null;
-    return isAlive;
-  });
-  // Array.sort sorts Number as if it was a String, 80 is before 9, use lodash.
-  return _.sortBy(workerIds);
-}
-
-function clusterSize() {
-  return liveWorkerIds().length;
-}
 
 var OPTION_DEFAULTS = {
   shutdownTimeout: 5000,
@@ -119,6 +105,8 @@ function _startDelay(self) {
 
 function resize(worker) {
   var self = this;
+
+  debug('_resize: w? %s', worker ? worker.id : undefined);
 
   _recordSuicide(self, worker);
