+++ package.json
@@ -19,8 +19,7 @@
   "main": "index.js",
   "scripts": {
     "pretest": "eslint --ignore-path .gitignore . && jscs .",
-    "test": "mocha --reporter spec",
-    "tap": "tap test/test-*.js"
+    "test": "tap --bail test/test-*.js"
   },
   "repository": {
     "type": "git",
@@ -37,8 +36,7 @@
   "devDependencies": {
     "eslint": "^1.1.0",
     "jscs": "^1.11.3",
-    "mocha": "^2.1.0",
-    "tap": "^0.7.1"
+    "tap": "^1.3.2"
   },
   "engines": {
     "node": "*"
